Admin and account functions

+++ routers/dbRoutes.py
@@ -8,14 +8,15 @@ from sqlmodel import Session, select
 
 #IMPORT FILES
 from core import security
-from schemas import dbSchema
+from schemas import dbSchema, authSchema
 from services import dbService, authService
 
 #IMPORT FILES VARIABLE BRIDGES
 vCoreSecurity = security
-vDbSchemas = dbSchema
+vDbSchema = dbSchema
 vDbService = dbService
 vAuthService = authService
+vAuthSchema = authSchema
 
 routerOwners = APIRouter(
     prefix="/owners",
@@ -27,15 +28,15 @@ routerUser = APIRouter(
     tags=["user"]
 )
 
-@routerOwners.post("/", response_model=vDbSchemas.ownerPublic)
-def createOwner(
+@routerOwners.post("/", response_model=vDbSchema.ownerPublic)
+def create_an_account(
     *,
     session: Session = Depends(vDbService.getSession),
-    owner: vDbSchemas.ownerCreate
+    owner: vDbSchema.ownerCreate
 ):
     
     #ASSURES UNIQUE USERNAME
-    existingOwner = session.query(vDbSchemas.owner).filter_by(username=owner.username).first()
+    existingOwner = session.query(vDbSchema.owner).filter_by(username=owner.username).first()
     if existingOwner:
         raise HTTPException(
             status_code=400,
@@ -45,7 +46,7 @@ def createOwner(
     #CREATES THE NEW USER
     hashedPassword = vCoreSecurity.hashPassword(owner.password)
     extraData = {"hashedPassword": hashedPassword}
-    dbOwner = vDbSchemas.owner.model_validate(owner, update=extraData)
+    dbOwner = vDbSchema.owner.model_validate(owner, update=extraData)
     session.add(dbOwner)
     session.commit()
     session.refresh(dbOwner)
@@ -55,56 +56,72 @@ def createOwner(
         
     return dbOwner
 
-@routerOwners.get("/", response_model=list[vDbSchemas.ownerPublic])
-def readOwners(
+@routerOwners.get("/", response_model=list[vDbSchema.ownerPublic])
+def fetch_all_accounts(
     *,
-    ownerPerm: vDbSchemas.owner = Depends(vAuthService.requireAnyPermission("admin")),
+    ownerPerm: vDbSchema.owner = Depends(vAuthService.requireAnyPermission("admin")),
     session: Session = Depends(vDbService.getSession),
     offset: int = 0,
-    limit: int = Query(default=100, le=100)
+    # limit: int = Query(default=100, le=100)
 ):
-    owners = session.exec(select(vDbSchemas.owner).offset(offset).limit(limit)).all()
+    owners = session.exec(select(vDbSchema.owner).offset(offset))
     return owners
 
-@routerOwners.get("/{ownerId}", response_model=vDbSchemas.ownerPublic)
-def readOwner(
+@routerOwners.get("/{ownerId}", response_model=vDbSchema.ownerPublic)
+def fetch_an_account(
     *,
-    ownerPerm: vDbSchemas.owner = Depends(vAuthService.requireAnyPermission("readOnly", "ownerPerm", "admin")),
+    ownerPerm: vDbSchema.owner = Depends(vAuthService.requireAnyPermission("readOnly", "ownerPerm", "admin")),
     session: Session = Depends(vDbService.getSession),
     ownerId: int,
-    currentUser: vDbSchemas.owner = Depends(vAuthService.getCurrentActiveUser)
+    currentUser: vDbSchema.owner = Depends(vAuthService.getCurrentActiveUser)
 ):
-    owner = session.get(vDbSchemas.owner, ownerId)
+    owner = session.get(vDbSchema.owner, ownerId)
+    permissionOwner = vAuthService.isAdmin(session, currentUser.ownerId)
+
     if not owner:
         raise HTTPException(status_code=404, detail="owner not found")
-    if owner.ownerId != currentUser.ownerId:
+    if not (owner.ownerId == currentUser.ownerId or permissionOwner):
         raise HTTPException(status_code=403, detail="Not authorized to access this owner")   
     
     return owner
 
-@routerUser.get("/me", response_model=vDbSchemas.ownerPublic)
-def getCurrentOwner(
+@routerUser.get("/me")
+def fetch_my_account(
     *,
-    currentUser: vDbSchemas.owner = Depends(vAuthService.getCurrentActiveUser),
-    ownerPerm: vDbSchemas.owner = Depends(vAuthService.requireAnyPermission("readOnly", "ownerPerm", "admin"))
+    session: Session = Depends(vDbService.getSession),
+    currentUser: vDbSchema.owner = Depends(vAuthService.getCurrentActiveUser)
 ):
-    return currentUser
+    
+    permissions = (
+        session.query(vAuthSchema.Permissions)
+        .filter_by(ownerId=currentUser.ownerId)
+        .first()
+    )
 
-@routerOwners.patch("/admin/{ownerId}", response_model=vDbSchemas.ownerPublic)
-def update_owner(
+    if not permissions:
+        raise HTTPException(status_code=404, detail="Permissions not found")
+
+    return {
+        "user": currentUser,
+        "permissions": {
+            "readOnly": permissions.readOnly,
+            "ownerPerm": permissions.ownerPerm,
+            "admin": permissions.admin}
+    }
+
+@routerOwners.patch("/admin/{ownerId}", response_model=vDbSchema.ownerPublic)
+def update_an_account(
     *,
-    ownerPerm: vDbSchemas.owner = Depends(vAuthService.requireAnyPermission("admin")),
+    ownerPerm: vDbSchema.owner = Depends(vAuthService.requireAnyPermission("admin")),
     session: Session = Depends(vDbService.getSession),
     ownerId: int,
-    owner: vDbSchemas.ownerUpdate,
-    currentUser: vDbSchemas.owner = Depends(vAuthService.getCurrentActiveUser)
+    owner: vDbSchema.ownerUpdate
 ):
     with Session(vDbService.engine) as session:
-       dbOwner = session.get(vDbSchemas.owner, ownerId)
+       dbOwner = session.get(vDbSchema.owner, ownerId)
     if not dbOwner:
         raise HTTPException(status_code=404, detail="owner not found")
-    if dbOwner.ownerId != currentUser.ownerId:
-        raise HTTPException(status_code=403, detail="Not authorized to access this owner")
+
     ownerData = owner.model_dump(exclude_unset=True)
     extraData = {}
     if "password" in ownerData:
@@ -117,13 +134,48 @@ def update_owner(
     session.refresh(dbOwner)
     return dbOwner
 
-@routerOwners.patch("/me", response_model=vDbSchemas.ownerPublic)
-def updateOwner(
+@routerOwners.get("/admin/{ownerId}")
+def read_an_account(
     *,
-    ownerPerm: vDbSchemas.owner = Depends(vAuthService.requireAnyPermission("readOnly", "ownerPerm", "admin")),
+    ownerPerm: vDbSchema.owner = Depends(vAuthService.requireAnyPermission("admin")),
     session: Session = Depends(vDbService.getSession),
-    owner: vDbSchemas.ownerUpdate,
-    currentUser: vDbSchemas.owner = Depends(vAuthService.getCurrentActiveUser)
+    ownerId: int
+):
+    owner = session.get(vDbSchema.owner, ownerId)
+
+    if not owner:
+        raise HTTPException(status_code=404, detail="owner not found")
+    
+    permissions = (
+        session.query(vAuthSchema.Permissions)
+        .filter_by(ownerId=owner.ownerId)
+        .first()
+    )
+
+    if not permissions:
+        raise HTTPException(status_code=404, detail="Permissions not found")
+    
+    if permissions and permissions.admin:
+        raise HTTPException(
+            status_code=403,
+            detail="You are not allowed to retreive an admin's full account dataset."
+        )
+
+    return {
+        "user": owner,
+        "permissions": {
+            "readOnly": permissions.readOnly,
+            "ownerPerm": permissions.ownerPerm,
+            "admin": permissions.admin}
+    }
+
+@routerUser.patch("/me", response_model=vDbSchema.ownerPublic)
+def update_my_account(
+    *,
+    ownerPerm: vDbSchema.owner = Depends(vAuthService.requireAnyPermission("readOnly", "ownerPerm", "admin")),
+    session: Session = Depends(vDbService.getSession),
+    owner: vDbSchema.ownerUpdate,
+    currentUser: vDbSchema.owner = Depends(vAuthService.getCurrentActiveUser)
 ):
     ownerData = owner.model_dump(exclude_unset=True)
     #ASSURES UNIQUE USERNAME
@@ -135,7 +187,7 @@ def updateOwner(
                 detail="This username is already your username"            
             )
 
-    existingOwner = session.query(vDbSchemas.owner).filter_by(username=owner.username).first()
+    existingOwner = session.query(vDbSchema.owner).filter_by(username=owner.username).first()
     if existingOwner:
         raise HTTPException(
             status_code=400,
@@ -143,7 +195,7 @@ def updateOwner(
         )
 
     with Session(vDbService.engine) as session:
-       dbOwner = session.get(vDbSchemas.owner, currentUser.ownerId)
+       dbOwner = session.get(vDbSchema.owner, currentUser.ownerId)
     if not dbOwner:
         raise HTTPException(status_code=404, detail="owner not found")
     if dbOwner.ownerId != currentUser.ownerId:
@@ -162,13 +214,42 @@ def updateOwner(
     return dbOwner
 
 @routerOwners.delete("/{ownerId}")
-def deleteOwner(
+def delete_an_owner(
     *,
-    ownerPerm: vDbSchemas.owner = Depends(vAuthService.requireAnyPermission("admin")),
+    ownerPerm: vDbSchema.owner = Depends(vAuthService.requireAnyPermission("admin")),
     session: Session = Depends(vDbService.getSession), ownerId: int,
-    currentUser: vDbSchemas.owner = Depends(vAuthService.getCurrentActiveUser)
+    currentUser: vDbSchema.owner = Depends(vAuthService.getCurrentActiveUser)
 ):
-    owner = session.get(vDbSchemas.owner, ownerId)
+    
+    # TO SEE IF OWNER EXISTS (BACKEND PROTECTION LOGIC)
+    owner = session.get(vDbSchema.owner, ownerId)
+    if not owner:
+        raise HTTPException(status_code=404, detail="owner not found")
+    
+    # TO STOP FELLOW ADMIN DELETE
+    targetPermissions = (
+        session.query(vAuthSchema.Permissions)
+        .filter_by(ownerId=ownerId)
+        .first()
+    )
+
+    if targetPermissions and targetPermissions.admin:
+        raise HTTPException(
+            status_code=403,
+            detail="You cannot delete a user with admin priviledges."
+        )
+
+    session.delete(owner)
+    session.commit()
+    return {"ok": True}
+
+@routerUser.delete("/me")
+def delete_my_account(
+    *,
+    session: Session = Depends(vDbService.getSession),
+    currentUser: vDbSchema.owner = Depends(vAuthService.getCurrentActiveUser)
+):
+    owner = session.get(vDbSchema.owner, currentUser.ownerId)
     if not owner:
         raise HTTPException(status_code=404, detail="owner not found")
     
@@ -176,17 +257,48 @@ def deleteOwner(
     session.commit()
     return {"ok": True}
 
-@routerUser.delete("/me")
-def deleteOwner(
+
+@routerOwners.delete("/{ownerId}")
+def delete_an_account(
     *,
-    ownerPerm: vDbSchemas.owner = Depends(vAuthService.requireAnyPermission("ownerPerm", "admin")),
-    session: Session = Depends(vDbService.getSession),
-    currentUser: vDbSchemas.owner = Depends(vAuthService.getCurrentActiveUser)
+    ownerPerm: vDbSchema.owner = Depends(vAuthService.requireAnyPermission("admin")),
+    session: Session = Depends(vDbService.getSession), ownerId: int,
+    currentUser: vDbSchema.owner = Depends(vAuthService.getCurrentActiveUser)
 ):
-    owner = session.get(vDbSchemas.owner, currentUser.ownerId)
+    
+    # TO SEE IF OWNER EXISTS (BACKEND PROTECTION LOGIC)
+    owner = session.get(vDbSchema.owner, ownerId)
     if not owner:
         raise HTTPException(status_code=404, detail="owner not found")
     
+    # TO STOP FELLOW ADMIN DELETE
+    targetPermissions = (
+        session.query(vAuthSchema.Permissions)
+        .filter_by(ownerId=ownerId)
+        .first()
+    )
+
+    if targetPermissions and targetPermissions.admin:
+        raise HTTPException(
+            status_code=403,
+            detail="You cannot delete a user with admin priviledges."
+        )
+
     session.delete(owner)
+    session.commit()
+    return {"ok": True}
+
+@routerUser.delete("/perm/me")
+def delete_my_permissions(
+    *,
+    ownerPerm: vDbSchema.owner = Depends(vAuthService.requireAnyPermission("ownerPerm", "admin")),
+    session: Session = Depends(vDbService.getSession),
+    currentUser: vDbSchema.owner = Depends(vAuthService.getCurrentActiveUser)
+):
+    targetPermissions = session.get(vAuthSchema.Permissions, currentUser.ownerId)
+    if not targetPermissions:
+        raise HTTPException(status_code=404, detail="owner not found")
+    
+    session.delete(targetPermissions)
     session.commit()
     return {"ok": True}
